Updated Zenodo links for data
The link to the tutorial data in Zenodo has been confirmed. In addition, all 8 links to individual datasets in Zenodo have been disabled using `http://<i></i>*.fastqsanger.gz`. These links were included so users could copy/paste them into the upload manager, not click on them (which results in the data being downloaded to the user's machine). I'm not sure if this method of disabling the links will break Travis :( If there is an alternate method, I'm all ears!