Fix workflow configuration & licensing concerns
The customization of the scatterplot steps that needs to be done by the
user in the workflow run dialog was described incompletely, which caused
problems in trainings when users selected column 4 both for plotting and
for grouping. The updated version explains in more detail the purpose of
the customization and clearly states which columns need to be selected
for plotting and grouping.

In addition, the "4Cs of diamonds" image got removed from the tutorial
(an is now only linked to) because of licensing concerns.

addresses issue: #1972