Update visualization with Circos traininig.
A comment block has been included to inform that it is necessary to use version 0.69.8+galaxy4 for the tutorial, since some of the parameters cannot be modified in other versions. In particular, from version 0.69.8+galaxy4 to 0.69.8+galaxy7 is not possible to select the cytogenetic bands and the option Convert bands from BED format to circos karyotype band format is not available.
In addition, the option Fill Band has been renamed as Band Transparency.