Replace method for generating HoF pages
This should be a performance improvement but it is hard to see initially
because previously they were registered as separate 'pages' and now it's
collected as a single page. Which does help identify potential
performance blockers though. And now we can e.g. centralise all of the
'find all tutorials for this author' step which was probably quite
slow.'

This also removes the need for the 'annoate-page' filter which was not
very useful in the first place'