Update tutorial.md
Changed line 244 from instructing the user to choose "Apply position restraints" to instead choosing "No position restraints". This is vital, and the without this, the simulation will crash. The reason is because, once the system is heated in NVT conditions and then goes into NPT, the atoms need to be allowed to move freely. If one restrains the atoms in NPT, and then takes those restraints off in the production run, said restrained atoms cannot suddenly adopt steady state velocities and the simulation crashes. 

It is important to note: Production run settings should generally, always be the same as those in the NPT runs. The only exception could be the number of steps (if the user wants a longer simulation for example), as well as the step size. Great tutorial though!

-BR