I feel dirty..
But wee need this link or some of the faqs will not be rendered properly on the page. This still mystifies me a bit but it seems we need to force jekyll to process stuff in a certain order for everything to work? So this filename needs to start with a bunch of a's.