Update tutorial.md
Figure 5 in this tutorial shows two genes each with two transcripts. gene 1 has UMI coloured pink and blue. Gene 2 has pink and green. The text, though, describes Gene 2 as being brown and green. This may be confusing to some students.
The commit here changes the tutorial text so that Gene two is described as pink and green.
The alternative would be to edit Figure 5?