Proposed changes to job-destinations and ansible-galaxy admin tutorials.
Following the job-destinations and ansible-galaxy instructions as they currently are, will both result in failed ansible playbook runs. See the details about the problems and my proposed fixed below. If my proposed fixes are not ideal, then feel free to reject them and just use this pull-request as informational messages about current problems in these tutorials:

* ansible-galaxy tutorial: Following this tutorial will result in ansible playbook errors at line 926 where the variable "miniconda_version: 4.12.0" is defined. This tutorial used to work without errors but recently the miniconda role fails when miniconda_version 4.12 is defined.
The cause of this is because the miniconda role by default attempts to install the latest conda version and then downgrades to the version defined in "miniconda_version". The latest miniconda versions can't be downgraded to 4.12 because a 'UnsatisfiableError' error will occur because of python version constraints. (the latest conda version being installed by the role uses python 3.11 which causes the version constraint error.) One workaround for this problem is to define extra variables, as can be seen in the diff. Two other possible solutions are to have the playbook install the latest conda version or to modify the code of the miniconda_role.

* job-destinations tutorial: In this tutorial, the student is instructed to run the ansible playbook after having added the usegalaxy_eu.tpv_auto_lint role. However running this playbook at this point will also result in ansible errors. This happens because TPV itself has not yet been installed. Skipping the playbook run at this point and only running it after the "Configuring TPV" section proved to be a workaround for me. As mentioned in the tutorial, adding TPV as a runner will cause Galaxy to automatically install TPV. The tpv_auto_lint role will only work at this point because it requires TPV to be installed. My proposed fix would be to remove the instruction to run the ansible-playbook at this point and run it after the "Configuring TPV" section.