Update Galaxy Interactive tools training (path-based tools and general review)
Include instructions to enable path-based interactive tools. Lift all sections one level up: it looks like the tutorial was intended to include a section to set up the Interactive Tools using Ansible and another to do it manually, but it includes only the former. As a result of this change, some warnings and comments had to have their location and/or type changed. Two questions, and @kysrpex as editor, were also added.